Add tests and improve exception handling (#2)
* Add initial tests and fix C-terminal mass

Implements basic tests for the pepfrag library, testing mass calculation and fragmentation functionality (loosely).

Also adds basic exception handling to cpepfrag to prevent crashes.

Note that tests for average mass calculation currently fail.

* Add test for many modifications

* Implement average mass calculation

* Fix Ion type following C++ changes

The return type from fragment is now a set of tuples, not the Ion namedtuple (for efficiency purposes). Therefore, introduct a type alias for Ion to be used in place of the namedtuple type.

* Improved exception handling for invalid residues

If an unnatural residue, such as U, is present in the peptide sequence, then calculate_mass will raise a python KeyError. Add tests for this behaviour.