Fix segmentation fault in with numpy types
Previously, if the modification was not a Python numeric type or a
string, the C++ code could segfault. This has been resolved by catching
this case and a test has been added for this. Numpy types are still not
supported in the modification site, but at least they don't cause a
segfault.